Fix integer overflow on 32-bit architectures
The gcc compiler claims a warning about integer overflow like this one:
   warning: integer overflow in expression of type 'long int' results in '-694967296' [-Woverflow]
for the e-collection-backend.c .

+++ src/addressbook/libedata-book/e-book-meta-backend.c
@@ -2214,7 +2214,7 @@ ebmb_notify_online_cb (GObject *object,
 		gint64 now = g_get_real_time ();
 
 		/* Do not auto-run refresh (due to getting online) more than once per hour */
-		if (now - meta_backend->priv->last_refresh_time >= G_USEC_PER_SEC * 60L * 60L) {
+		if (now - meta_backend->priv->last_refresh_time >= G_USEC_PER_SEC * ((gint64) 60) * 60) {
 			meta_backend->priv->last_refresh_time = now;
 
 			e_book_meta_backend_schedule_refresh (meta_backend);
