+++ src/calendar/libedata-cal/e-cal-meta-backend.c
@@ -3109,7 +3109,7 @@ ecmb_notify_online_cb (GObject *object,
 		gint64 now = g_get_real_time ();
 
 		/* Do not auto-run refresh (due to getting online) more than once per hour */
-		if (now - meta_backend->priv->last_refresh_time >= G_USEC_PER_SEC * 60L * 60L) {
+		if (now - meta_backend->priv->last_refresh_time >= G_USEC_PER_SEC * ((gint64) 60) * 60) {
 			meta_backend->priv->last_refresh_time = now;
 
 			e_cal_meta_backend_schedule_refresh (meta_backend);
